# Consistently disable the garbage collector in Bootstrap.php