--- a/src/phan.php
+++ b/src/phan.php
@@ -2,11 +2,6 @@
 
 declare(strict_types=1);
 
-// Phan does a ton of GC and this offers a major speed
-// improvement if your system can handle it (which it
-// should be able to)
-gc_disable();
-
 // Check the environment to make sure Phan can run successfully
 require_once(__DIR__ . '/requirements.php');