--- a/src/prep.php
+++ b/src/prep.php
@@ -2,11 +2,6 @@
 
 declare(strict_types=1);
 
-// Phan does a ton of GC and this offers a major speed
-// improvement if your system can handle it (which it
-// should be able to)
-gc_disable();
-
 // Check the environment to make sure Phan can run successfully
 // @phan-file-suppress PhanPluginRemoveDebugEcho
 require_once __DIR__ . '/requirements.php';